Added basic index page, static files directory, and nginx webserver
Created index.html which will server as the one and only page for the
site.  Also modified server.py to serve it instead of nginx in case it
needs to be dynamic later.

The nginx server acts as a reverse proxy for all incoming trafic on port
80 (except for static directories) and redirects it to port 5000
(the flask app).  The static directories are currently /scripts/ and
/css/ for javascript and css files respectively.  They're rooted at
/home/adin/myles/static.

NOTE: The root of the static locations needs to be changed in
production to where the static files are rooted.