Core functionality flushed out.  Video player, commpression, and compress button working.
A commit is long overdue.  Got so caught up jumping from feature to
feature that there was never a place that felt right for a commit.

The baseline for everything is done.  A video plays in the browser,
served statically by nginx.  A compress button on the webpage posts a
message to the server telling it to compress the video which it does in
a seperate thread.  Video playback isn't interrupted until the
compression is complete.

There's still a bunch to be done, but at this point the app is
functionally complete.