The javascript now disables the compress button if the server is compressing at page load time or if it is pressed once
The onLoad() javascript function is run when the file is loaded in the
header.  It checks if the server is compressing and both reports the
times the movie has been compressed and disables the compression button
if necessary.

The compress button styling has also been commented out because it made
it impossible to tell if it was disabled.