Switched to gunicorn for the wsgi server in preperation for server side events
The configuration for gunicorn is in gunicorn.conf.py and is pretty self
explanitory.  It sets up a bunch of logging and then chooses the worker
type and number of threads.  For this I'm using gevent workers so that
sse streams don't block other incoming http requests.  This is what was
recommended by flask_sse.  This broke the current compression state
system, but that's going to be replaced with redis since it's already
needed as a backend for the sse.

Also added a fresh clone script and a readme to roughly list the non-pip
requirements.  This might be moved in the future.

Finally, I added a network.json file which will list things like the
redis server ip and password, as well as any other ips needed for the
system.  It is marked skip-worktree so the sensitive information should
never be pushed.  It also makes moving from the dev environment to the
deployment environment easier.